requirements.txt
streamlit
pandas
openpyxl
plotly

PyYAML

pydantic
